fix: dupplication code in testing, so the 2 tests are merged
